% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicefData.R
\name{add_country_metadata}
\alias{add_country_metadata}
\title{Add country-level metadata columns}
\usage{
add_country_metadata(df, metadata_list)
}
\description{
Add country-level metadata columns
}
\keyword{internal}
