% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicefData.R
\name{add_indicator_metadata}
\alias{add_indicator_metadata}
\title{Add indicator-level metadata columns}
\usage{
add_indicator_metadata(df, metadata_list)
}
\description{
Add indicator-level metadata columns
}
\keyword{internal}
