% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_generic.R
\name{clean_generic}
\alias{clean_generic}
\title{Tidy up generic data sets.}
\usage{
clean_generic(filename)
}
\arguments{
\item{filename}{path to source file}
}
\value{
a tibble of tidy data
}
\description{
Tidy up generic data sets.
}
