% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_infant_hiv.R
\name{clean_infant_hiv}
\alias{clean_infant_hiv}
\title{Tidy up the infant HIV data set.}
\usage{
clean_infant_hiv(filename)
}
\arguments{
\item{filename}{path to source file}
}
\value{
a tibble of tidy data
}
\description{
Tidy up the infant HIV data set.
}
