% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{compute_data_hash}
\alias{compute_data_hash}
\title{Compute hash of data frame for version tracking}
\usage{
compute_data_hash(df)
}
\arguments{
\item{df}{Data frame to hash}
}
\value{
Character hash string (16 characters)
}
\description{
Compute hash of data frame for version tracking
}
