% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_registry.R
\name{.fetch_codelist}
\alias{.fetch_codelist}
\title{Fetch indicator codelist from UNICEF SDMX API}
\usage{
.fetch_codelist(codelist_id)
}
\value{
Named list of indicator metadata
}
\description{
Fetch indicator codelist from UNICEF SDMX API
}
\keyword{internal}
