% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{.get_basic_dataflow_info}
\alias{.get_basic_dataflow_info}
\title{Get basic dataflow info from _unicefdata_dataflows.yaml}
\usage{
.get_basic_dataflow_info(dataflow, metadata_dir)
}
\description{
Get basic dataflow info from _unicefdata_dataflows.yaml
}
\keyword{internal}
