% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_cache.R
\name{.get_cached_schema}
\alias{.get_cached_schema}
\title{Retrieve Schema from Cache}
\usage{
.get_cached_schema(key, namespace = "schema")
}
\arguments{
\item{key}{Cache key}

\item{namespace}{Cache namespace}
}
\value{
Schema object if found, NULL otherwise
}
\description{
Retrieve Schema from Cache
}
\keyword{internal}
