% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_sync.R
\name{.get_metadata_dir}
\alias{.get_metadata_dir}
\title{Get metadata directory path}
\usage{
.get_metadata_dir()
}
\value{
Path to metadata directory: prefers \code{inst/metadata/current} in the source tree, otherwise a user cache directory at \code{tools::R_user_dir("unicefData", "cache")/metadata/current}.
}
\description{
Get metadata directory path
}
\keyword{internal}
