% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_cache.R
\name{.is_cached}
\alias{.is_cached}
\title{Check if Schema is Cached}
\usage{
.is_cached(key, namespace = "schema")
}
\arguments{
\item{key}{Cache key}

\item{namespace}{Cache namespace}
}
\value{
Logical TRUE if schema is cached
}
\description{
Check if Schema is Cached
}
\keyword{internal}
