% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicef_core.R
\name{.is_http_404}
\alias{.is_http_404}
\title{Check if an error is an HTTP 404 response}
\usage{
.is_http_404(e)
}
\arguments{
\item{e}{An error condition object}
}
\value{
TRUE if the error represents a 404 Not Found, FALSE otherwise
}
\description{
Check if an error is an HTTP 404 response
}
\keyword{internal}
