% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{.load_yaml}
\alias{.load_yaml}
\title{Load YAML file from current metadata directory}
\usage{
.load_yaml(filename)
}
\arguments{
\item{filename}{Character string of filename (without path)}
}
\value{
List from parsed YAML, or empty list if file not found
}
\description{
Load YAML file from current metadata directory
}
\keyword{internal}
