% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{.load_yaml_from_path}
\alias{.load_yaml_from_path}
\title{Load YAML file from absolute path}
\usage{
.load_yaml_from_path(filepath)
}
\arguments{
\item{filepath}{Character string of absolute file path}
}
\value{
List from parsed YAML, or empty list if file not found
}
\description{
Load YAML file from absolute path
}
\keyword{internal}
