% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_registry.R
\name{.parse_codelist_xml}
\alias{.parse_codelist_xml}
\title{Parse SDMX codelist XML response}
\usage{
.parse_codelist_xml(xml_content)
}
\arguments{
\item{xml_content}{Character. Raw XML content from API}
}
\value{
Named list of indicator metadata
}
\description{
Parse SDMX codelist XML response
}
\keyword{internal}
