% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_registry.R
\name{.save_cache}
\alias{.save_cache}
\title{Save indicator metadata to cache file}
\usage{
.save_cache(indicators)
}
\arguments{
\item{indicators}{Named list of indicator metadata}
}
\description{
Save indicator metadata to cache file
}
\keyword{internal}
