% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_registry.R, R/metadata.R
\name{.yaml_no_wrap}
\alias{.yaml_no_wrap}
\title{Convert R list to YAML without line wrapping}
\usage{
.yaml_no_wrap(data, indent = 0)

.yaml_no_wrap(data, indent = 0)
}
\arguments{
\item{data}{List to convert}

\item{indent}{Current indentation level}
}
\value{
Character vector of YAML lines

Character vector of YAML lines
}
\description{
Convert R list to YAML without line wrapping

Convert R list to YAML without line wrapping
}
\keyword{internal}
