% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_registry.R, R/metadata.R
\name{.yaml_scalar}
\alias{.yaml_scalar}
\title{Convert scalar value to YAML string}
\usage{
.yaml_scalar(x)

.yaml_scalar(x)
}
\arguments{
\item{x}{Scalar value}
}
\value{
Character string in YAML format

Character string in YAML format
}
\description{
Convert scalar value to YAML string

Convert scalar value to YAML string
}
\keyword{internal}
