% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unicef.R
\name{fetch_sdmx}
\alias{fetch_sdmx}
\title{Fetch SDMX content from URL}
\usage{
fetch_sdmx(url, ua, retry)
}
\arguments{
\item{url}{URL to fetch}

\item{ua}{User agent string}

\item{retry}{Number of retries}
}
\value{
Content as text
}
\description{
Fetch SDMX content from URL
}
\keyword{internal}
