% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicef_core.R
\name{fetch_sdmx_text}
\alias{fetch_sdmx_text}
\title{Fetch SDMX content as text}
\usage{
fetch_sdmx_text(url, ua = .unicefData_ua, retry)
}
\arguments{
\item{url}{URL to fetch}

\item{ua}{User agent string}

\item{retry}{Number of retries}
}
\value{
Content as text
}
\description{
Fetch SDMX content as text
}
\keyword{internal}
