% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_codelist_meta}
\alias{get_codelist_meta}
\title{Get metadata for a specific codelist}
\usage{
get_codelist_meta(codelist_id)
}
\arguments{
\item{codelist_id}{Codelist identifier}
}
\value{
List with codelist metadata or NULL
}
\description{
Get metadata for a specific codelist
}
