% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_loader.R
\name{get_config_path}
\alias{get_config_path}
\title{Get path to the shared indicators.yaml config file}
\usage{
get_config_path()
}
\value{
Path to indicators.yaml
}
\description{
Searches in order:
\enumerate{
\item UNICEF_CONFIG_PATH environment variable
\item ../../config/indicators.yaml relative to this file
\item ./config/indicators.yaml relative to current working directory
}
}
