% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_sync.R
\name{get_dataflow_list}
\alias{get_dataflow_list}
\title{Get list of all UNICEF dataflows}
\usage{
get_dataflow_list(max_retries = 3)
}
\arguments{
\item{max_retries}{Number of retries}
}
\value{
Tibble with dataflow info
}
\description{
Get list of all UNICEF dataflows
}
