% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_dataflow_meta}
\alias{get_dataflow_meta}
\title{Get metadata for a specific dataflow}
\usage{
get_dataflow_meta(dataflow_id)
}
\arguments{
\item{dataflow_id}{Dataflow identifier}
}
\value{
List with dataflow metadata or NULL
}
\description{
Get metadata for a specific dataflow
}
