% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_sync.R
\name{get_dataflow_schema}
\alias{get_dataflow_schema}
\title{Get schema for a specific dataflow}
\usage{
get_dataflow_schema(dataflow_id, version = "1.0", max_retries = 3)
}
\arguments{
\item{dataflow_id}{Dataflow ID}

\item{version}{Dataflow version}

\item{max_retries}{Number of retries}
}
\value{
List with dimensions, attributes, etc. or NULL
}
\description{
Get schema for a specific dataflow
}
