% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_sync.R
\name{get_expected_columns}
\alias{get_expected_columns}
\title{Get list of expected column names for a dataflow}
\usage{
get_expected_columns(dataflow_id, metadata_dir = NULL)
}
\arguments{
\item{dataflow_id}{Dataflow ID (e.g., 'CME', 'NUTRITION')}

\item{metadata_dir}{Directory containing dataflow_schemas.yaml}
}
\value{
Character vector of column names (dimensions + time + attributes)
}
\description{
Get list of expected column names for a dataflow
}
