% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicefData.R
\name{get_income_groups}
\alias{get_income_groups}
\title{Get ISO3 to World Bank income group mapping}
\usage{
get_income_groups()
}
\description{
Get ISO3 to World Bank income group mapping
}
\keyword{internal}
