% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_indicator_meta}
\alias{get_indicator_meta}
\title{Get metadata for a specific indicator}
\usage{
get_indicator_meta(indicator_code)
}
\arguments{
\item{indicator_code}{Indicator code}
}
\value{
List with indicator metadata or NULL
}
\description{
Get metadata for a specific indicator
}
