% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_loader.R
\name{get_indicators_by_sdg}
\alias{get_indicators_by_sdg}
\title{Get indicator codes by SDG goal}
\usage{
get_indicators_by_sdg(sdg_goal, config_path = NULL)
}
\arguments{
\item{sdg_goal}{SDG goal number (e.g., '3', '4')}

\item{config_path}{Optional explicit path to config file}
}
\value{
Character vector of indicator codes
}
\description{
Get indicator codes by SDG goal
}
