% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicef_core.R
\name{get_unicef_raw}
\alias{get_unicef_raw}
\title{Fetch Raw UNICEF Data}
\usage{
get_unicef_raw(
  indicator = NULL,
  dataflow = NULL,
  countries = NULL,
  start_year = NULL,
  end_year = NULL,
  max_retries = 3,
  version = NULL,
  page_size = 1e+05,
  verbose = TRUE
)
}
\description{
Low-level fetcher for UNICEF SDMX API.
}
