% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infant_hiv.R
\docType{data}
\name{infant_hiv}
\alias{infant_hiv}
\title{Tidied infant HIV data.}
\format{A data frame
\describe{
  \item{country}{Country reporting (ISO3 code)}
  \item{kind}{Type of report (low, estimate, high)}
  \item{reported}{Value reported (may be NA)}
  \item{year}{Year reported}
}}
\usage{
infant_hiv
}
\description{
This tidy data is derived from the `infant_hiv.csv` file, which in turn is
derived from an Excel spreadsheet provided by UNICEF - see the README.md file
in the raw data directory for details.
}
\keyword{datasets}
