% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unicef.R
\name{list_unicef_codelist}
\alias{list_unicef_codelist}
\title{List SDMX codelist for a given flow + dimension}
\usage{
list_unicef_codelist(
  flow,
  dimension,
  cache_dir = tools::R_user_dir("get_unicef", "cache"),
  retry = 3
)
}
\arguments{
\item{flow}{character flow ID, e.g. "NUTRITION"}

\item{dimension}{character dimension ID within that flow, e.g. "INDICATOR"}
}
\value{
A tibble with columns \code{code} and \code{description}
}
\description{
List SDMX codelist for a given flow + dimension
}
