% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unicef.R
\name{list_unicef_flows}
\alias{list_unicef_flows}
\title{List available UNICEF SDMX “flows”}
\usage{
list_unicef_flows(
  cache_dir = tools::R_user_dir("get_unicef", "cache"),
  retry = 3
)
}
\value{
A tibble with columns \code{id}, \code{agency}, and \code{version}
}
\description{
Download and cache the SDMX data-flow definitions from UNICEF’s REST endpoint.
}
