% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{load_codelists}
\alias{load_codelists}
\title{Load cached codelist metadata from YAML}
\usage{
load_codelists()
}
\value{
List with codelist metadata
}
\description{
Load cached codelist metadata from YAML
}
