% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{load_dataflows}
\alias{load_dataflows}
\title{Load cached dataflow metadata from YAML}
\usage{
load_dataflows()
}
\value{
List with dataflow metadata
}
\description{
Load cached dataflow metadata from YAML
}
