% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{load_indicators}
\alias{load_indicators}
\title{Load cached indicator metadata from YAML}
\usage{
load_indicators()
}
\value{
List with indicator metadata
}
\description{
Load cached indicator metadata from YAML
}
