% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{load_sync_summary}
\alias{load_sync_summary}
\title{Load last sync summary (deprecated, use load_sync_history)}
\usage{
load_sync_summary()
}
\value{
List with sync summary from latest vintage
}
\description{
Load last sync summary (deprecated, use load_sync_history)
}
