% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{print.unicef_dataflow_schema}
\alias{print.unicef_dataflow_schema}
\title{Print method for dataflow schema}
\usage{
\method{print}{unicef_dataflow_schema}(x, ...)
}
\arguments{
\item{x}{A unicef_dataflow_schema object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for dataflow schema
}
