% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{sync_codelists}
\alias{sync_codelists}
\title{Sync codelist definitions from SDMX API}
\usage{
sync_codelists(codelist_ids = NULL, verbose = TRUE)
}
\arguments{
\item{codelist_ids}{Vector of codelist IDs to sync (default: common codelists)}

\item{verbose}{Print progress messages}
}
\value{
List with codelist metadata
}
\description{
Sync codelist definitions from SDMX API
}
