% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{sync_dataflows}
\alias{sync_dataflows}
\title{Sync dataflow definitions from SDMX API}
\usage{
sync_dataflows(verbose = TRUE)
}
\arguments{
\item{verbose}{Print progress messages}
}
\value{
List with dataflow metadata
}
\description{
Sync dataflow definitions from SDMX API
}
