% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_registry.R, R/metadata_sync.R
\name{.fetch_codelist}
\alias{.fetch_codelist}
\title{Fetch indicator codelist from UNICEF SDMX API}
\usage{
.fetch_codelist(codelist_id)

.fetch_codelist(codelist_id)
}
\arguments{
\item{codelist_id}{Codelist ID}
}
\value{
Named list of indicator metadata

List with codes and metadata
}
\description{
Fetch indicator codelist from UNICEF SDMX API

Fetch a single codelist
}
\keyword{internal}
