% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_sync.R
\name{.fetch_codelist}
\alias{.fetch_codelist}
\title{Fetch a single codelist}
\usage{
.fetch_codelist(codelist_id)
}
\arguments{
\item{codelist_id}{Codelist ID}
}
\value{
List with codes and metadata
}
\description{
Fetch a single codelist
}
\keyword{internal}
