% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_sync.R
\name{.get_metadata_dir}
\alias{.get_metadata_dir}
\title{Get metadata directory path}
\usage{
.get_metadata_dir()
}
\value{
Path to R/metadata/current/
}
\description{
Get metadata directory path
}
\keyword{internal}
