% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_registry.R
\name{.infer_category}
\alias{.infer_category}
\title{Infer dataflow category from indicator code prefix}
\usage{
.infer_category(indicator_code)
}
\arguments{
\item{indicator_code}{Character. The indicator code}
}
\value{
Character. The inferred dataflow name
}
\description{
Infer dataflow category from indicator code prefix
}
\keyword{internal}
