% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_sync.R
\name{.save_yaml}
\alias{.save_yaml}
\title{Save YAML with proper formatting}
\usage{
.save_yaml(filename, data, output_dir)
}
\arguments{
\item{filename}{Filename}

\item{data}{Data to save}

\item{output_dir}{Output directory}
}
\description{
Save YAML with proper formatting
}
\keyword{internal}
