% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicef_core.R
\name{filter_unicef_data}
\alias{filter_unicef_data}
\title{Filter UNICEF Data (Sex, Age, Wealth, etc.)}
\usage{
filter_unicef_data(
  df,
  sex = NULL,
  age = NULL,
  wealth = NULL,
  residence = NULL,
  maternal_edu = NULL,
  verbose = TRUE
)
}
\description{
Filters data to specific disaggregations or defaults to totals.
}
