% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_registry.R
\name{get_cache_info}
\alias{get_cache_info}
\title{Get Cache Info}
\usage{
get_cache_info()
}
\value{
Named list with cache metadata
}
\description{
Get information about the current cache state.
}
\examples{
\dontrun{
info <- get_cache_info()
print(info$cache_path)
print(info$indicator_count)
}

}
