% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicefData.R
\name{list_unicef_codelist}
\alias{list_unicef_codelist}
\title{List SDMX codelist for a given flow + dimension}
\usage{
list_unicef_codelist(
  flow,
  dimension,
  cache_dir = tools::R_user_dir("unicefData", "cache"),
  retry = 3
)
}
\arguments{
\item{flow}{character flow ID, e.g. "NUTRITION"}

\item{dimension}{character dimension ID within that flow, e.g. "INDICATOR"}

\item{cache_dir}{Character path to cache directory.}

\item{retry}{Integer, number of retries.}
}
\value{
A tibble with columns \code{code} and \code{description}
}
\description{
List SDMX codelist for a given flow + dimension
}
