% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicefData.R
\name{list_unicef_flows}
\alias{list_unicef_flows}
\title{List Available UNICEF SDMX Flows}
\usage{
list_unicef_flows(
  cache_dir = tools::R_user_dir("unicefData", "cache"),
  retry = 3
)
}
\arguments{
\item{cache_dir}{Character path to cache directory.}

\item{retry}{Integer, number of retries.}
}
\value{
A tibble with columns \code{id}, \code{agency}, and \code{version}
}
\description{
Download and cache the SDMX data-flow definitions from the UNICEF REST endpoint.
}
