% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_registry.R
\name{refresh_indicator_cache}
\alias{refresh_indicator_cache}
\title{Refresh Indicator Cache}
\usage{
refresh_indicator_cache()
}
\value{
Integer. Number of indicators in the refreshed cache
}
\description{
Force refresh of the indicator cache from UNICEF SDMX API.
}
\examples{
\dontrun{
n <- refresh_indicator_cache()
message(sprintf("Refreshed cache with \%d indicators", n))
}

}
