% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{set_metadata_cache}
\alias{set_metadata_cache}
\title{Set metadata cache directory}
\usage{
set_metadata_cache(path = NULL)
}
\arguments{
\item{path}{Path to cache directory}
}
\description{
Set metadata cache directory
}
