% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R, R/metadata_sync.R
\name{sync_codelists}
\alias{sync_codelists}
\title{Sync codelist definitions from SDMX API (excluding countries/regions)}
\usage{
sync_codelists(codelist_ids = NULL, verbose = TRUE, output_dir = NULL)

sync_codelists(codelist_ids = NULL, verbose = TRUE, output_dir = NULL)
}
\arguments{
\item{codelist_ids}{Vector of codelist IDs}

\item{verbose}{Print progress}

\item{output_dir}{Output directory}
}
\value{
List with codelist metadata

List of codelists
}
\description{
Sync codelist definitions from SDMX API (excluding countries/regions)

Sync codelists (excluding countries/regions)
}
