% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R, R/metadata_sync.R
\name{sync_dataflows}
\alias{sync_dataflows}
\title{Sync dataflow definitions from SDMX API}
\usage{
sync_dataflows(verbose = TRUE, output_dir = NULL)

sync_dataflows(verbose = TRUE, output_dir = NULL)
}
\arguments{
\item{verbose}{Print progress}

\item{output_dir}{Output directory}
}
\value{
List with dataflow metadata

List of dataflows
}
\description{
Sync dataflow definitions from SDMX API

Sync dataflow definitions
}
