% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicefdata.R
\docType{package}
\name{unicefdata}
\alias{unicefdata}
\alias{unicefdata-package}
\title{Clean up and share some data from UNICEF on infant HIV rates.}
\description{
Clean up and share some data from UNICEF on infant HIV rates.
}
\author{
Greg Wilson, \email{gvwilson@third-bit.com}
}
